fix: 密码放到面板设置界面
